requirements.txt
# Core dependencies
fastapi==0.100.0
uvicorn[standard]==0.23.2
sqlalchemy==2.0.20
python-jose[cryptography]==3.4.0
bcrypt==4.3.0
python-multipart==0.0.19
email-validator==2.0.0
python-dotenv==1.0.0

# Database
alembic==1.11.2

# Development tools (optional)
httpx==0.24.1  # For TestClient
pydantic~=2.11.7
pytest~=7.4.0
factory-boy~=3.3.0
Faker~=19.3.0
requests~=2.32.4